+++ NEWS.md
@@ -19,7 +19,7 @@
  
  - Fix incorrect information message when observations are removed because of infinite values (in some circumstances the removal was wrongly attributed to NAness).
  
- - Fix bug in `etable` when checking the argument `coefstat`.
+ - Fix bug in `etable` when checking the argument `coefstat` (thanks to @waynelapierre [#121](https://github.com/lrberge/fixest/issues/121)).
 
 ## etable